# Conflicts:
#	common/src/main/java/com/stevekung/fishofthieves/compatibility/biolith/FOTSurfaceRuleData.java
#	common/src/main/java/com/stevekung/fishofthieves/entity/shoal/ShoalFishData.java
#	common/src/main/java/com/stevekung/fishofthieves/feature/TropicalIslandBlockBlobFeature.java
#	common/src/main/java/com/stevekung/fishofthieves/feature/foliageplacers/CoconutFrondsPlacer.java
#	common/src/main/java/com/stevekung/fishofthieves/feature/surfacerules/BlockStateConditionSource.java
#	common/src/main/java/com/stevekung/fishofthieves/feature/surfacerules/WaterSurroundedConditionSource.java
#	common/src/main/java/com/stevekung/fishofthieves/feature/trunkplacers/CoconutTrunkPlacer.java
#	common/src/main/java/com/stevekung/fishofthieves/loot/FOTLootManager.java
#	common/src/main/java/com/stevekung/fishofthieves/network/SyncClientShoalFishPacket.java
#	fabric/src/gametest/java/com/stevekung/fishofthieves/fabric/gametest/ShoalTestSuite.java
#	fabric/src/main/java/com/stevekung/fishofthieves/fabric/datagen/provider/BlockLootProvider.java
#	fabric/src/main/resources/fish_of_thieves.accesswidener
#	gradle.properties

+++ common/src/main/java/com/stevekung/fishofthieves/loot/condition/FOTLocationCheck.java
@@ -29,7 +29,7 @@ public record FOTLocationCheck(Optional<FOTLocationPredicate> predicate, BlockPo
     {
         var vec3 = context.getOptionalParameter(LootContextParams.ORIGIN);
         var entity = context.getOptionalParameter(LootContextParams.THIS_ENTITY);
-        return vec3 != null && (this.predicate.isEmpty() || this.predicate.get().matches(context.getLevel(), entity, vec3.x() + (double) this.offset.getX(), vec3.y() + (double) this.offset.getY(), vec3.z() + (double) this.offset.getZ()));
+        return vec3 != null && (this.predicate.isEmpty() || this.predicate.get().matches(context.getLevel(), entity, vec3.x() + this.offset.getX(), vec3.y() + this.offset.getY(), vec3.z() + this.offset.getZ()));
     }
 
     public static Builder checkLocation(FOTLocationPredicate.Builder locationPredicateBuilder)
